requirements.txt
# requirements.txt

fastapi>=0.115.5
uvicorn[standard]>=0.32.1
temporalio>=1.7.1
pydantic>=2.12.2
python-dotenv>=1.0.1

# Database
psycopg[binary]==3.2.3
psycopg-pool==3.2.3
# psycopg-pool manages a pool of reusable Postgres connections so we don't
# open a new connection for every query.

# Repo cloning
gitpython==3.1.43
# Python wrapper around git. Lets us run "git clone" from Python code.

# Code parsing / chunking
tree-sitter==0.23.2
tree-sitter-python==0.23.4
# tree-sitter parses source code into a syntax tree so we can extract
# functions and classes as individual chunks instead of splitting by line count.
# tree-sitter-python is the language pack for Python files.
# Add more language packs (tree-sitter-javascript etc.) later for broader support.

# OpenAI Agents SDK — multi-agent framework (pulls in openai>=2.26 as a dependency)
openai-agents>=0.13.0
# Used for:
#   - Explorer agent with file tools (M3)
#   - Router + specialist handoffs (M5+)
